requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.4.17
chardet==3.0.4
Django==1.9.8
django-crispy-forms==1.6.1
django-formtools==2.0
django-pure-pagination==0.3.0
django-simple-captcha==0.4.6
DjangoUeditor==1.8.143
future==0.16.0
httplib2==0.9.2
idna==2.5
MySQL-python==1.2.5
olefile==0.44
Pillow==4.1.1
PyMySQL==0.7.11
requests==2.18.1
six==1.10.0
urllib3==1.21.1abeautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.4.17
chardet==3.0.4
Django==1.9.8
django-crispy-forms==1.6.1
django-formtools==2.0
django-pure-pagination==0.3.0
django-simple-captcha==0.4.6
DjangoUeditor==1.8.143
future==0.16.0
httplib2==0.9.2
idna==2.5
Markdown==2.6.8
MySQL-python==1.2.5
olefile==0.44
Pillow==4.1.1
Pygments==2.2.0
PyMySQL==0.7.11
requests==2.18.1
six==1.10.0
urllib3==1.21.1
xlrd==1.0.0
xlrd==1.0.0
